+++ app/Models/VoucherTransaction.php
@@ -558,7 +558,7 @@ class VoucherTransaction extends Model
             return null;
         }
 
-        return max(now()->diffInDays($this->transfer_at), 0);
+        return max((int) now()->diffInDays($this->transfer_at), 0);
     }
 
     /**
